Initial cleanup of script.
Second option is now an identifier which allows for a customized "runno," instead of automatically generating one from the name of the input nii4D.
Third option is flag for dti data, if an attempt to extract a btable from a .bxh header is desired.

Cleaned up typo that was generating a folder named '}' in the '-results' folder.